requirements.txt
kornia==0.3.0
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.5.1.48
Pillow==8.1.0
PyYAML==5.4.1
scikit-learn==0.23.2
scipy==1.5.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.55.1
